src/django/requirements.txt
psycopg2-binary
django
djangorestframework
djangorestframework-simplejwt
hvac